Move config-from-json to common config module
